refactor(runtime): split host.rs by concern
`host.rs` was the largest production file in the workspace at 2,063 lines,
holding three things that only shared a file: the vocabulary clients speak, the
event stream, and the host that interprets both.

- `host/types.rs` (430) - `SpawnArgs`, `RunListEntry`, `RunListing`,
  `DaemonHealth`, the five callback aliases, `SubAgentOp`, `SubAgentReport`,
  `ControlOp`. These are the crate's public vocabulary; the host is just the
  thing that happens to interpret them.
- `host/events.rs` (216) - `WorldEvent` and its impl, `WorldEventSink`,
  `status_str`, and the `Emitted` snapshot. Kept together because the snapshot
  exists to decide what is worth emitting, so the two only make sense as a pair.
- `host/mod.rs` (1,459) - `WorldHost`, its constants, and its impl.

Glob re-exported per `pipeline/mod.rs`, so every existing `host::ControlOp` and
`host::WorldEvent` path is untouched and no call site changed.

**Verified pure rather than asserted pure.** Sorting every non-comment,
non-import line before and after gives 1,215 in and 1,214 out, and the whole
difference is accounted for: 13 items (`Emitted`, its ten fields, `status_str`)
gained `pub(super)` because they now live in a submodule, `#[path]` moved one
directory deeper, and a duplicated `#[cfg(test)]` attribute - two of them stacked
on the same item - is gone. That is the missing line.

`host_tests.rs` now names `oneshot` and the interaction types itself instead of
inheriting them through `super::*`. The host no longer mentions either, so
leaning on the parent would have meant re-adding two imports production does not
use, which `-D warnings` rejects.

1,155 runtime tests pass and the package holds 100%.
